update build (#3)
* update build

* tty formatting

+++ .travis.sh
@@ -1,9 +1,17 @@
 #!/bin/sh
 set -e
 
+#go get -v github.com/golang/lint/golint
+#[ "0" = "$?" ] || exit 1
+
+go get -u github.com/golang/dep/cmd/dep
+[ "0" = "$?" ] || exit 2
+
+dep ensure
+[ "0" = "$?" ] || exit 4
+
 rm -f coverage.txt
-for dir in $(go list ./...); do
-    echo "go test -timeout 20s -coverprofile=profile.out $dir"
+for dir in $(go list ./... | grep -v vendor); do
     go test -timeout 20s -coverprofile=profile.out $dir
     exit_code=$?
     if [ "0" != "$exit_code" ]; then
@@ -14,4 +22,15 @@ for dir in $(go list ./...); do
         rm profile.out
     fi
 done
+
+#exit_code=0
+#for dir in $(go list ./... | grep -v vendor); do
+#    echo "golint $dir"
+#    result=$(golint $dir)
+#    if [ "" != "$result" ]; then
+#        echo $result
+#        exit_code=5
+#    fi
+#done
+
 exit $exit_code
